Add cleanup method

+++ lib/raspi-gpio.rb
@@ -1,10 +1,15 @@
-# raspi-gpio - A simple and light interface to interact with GPIO pins of the Raspberry Pi.
-#   Guide and source code : https://github.com/exybore/raspi-gpio-rb
-#   Documentation : https://rubydoc.info/gems/raspi-gpio
+# A simple and light interface to interact with GPIO pins of the Raspberry Pi.
+#   Guide and source code : https://github.com/sergio1990/raspi-gpio-rb
 #
-# @author exybore
-
-# Main class, to manage GPIO pins
+# @author exybore, sergio1990
+#
+# @example Common usage
+#   pin = GPIO.new(12, GPIO::OUT)
+#   pin.value = GPIO::LOW
+#   pin.mode = GPIO::IN
+#   pin.value # => either LOW or HIGH
+#   pin.cleanup
+#
 class GPIO
   # GPIO core library path
   LIB_PATH = '/sys/class/gpio'
@@ -49,6 +54,11 @@ class GPIO
   class BadValueError < BaseError
   end
 
+  # Exception to handle cases when trying to work with the pin which was
+  # already cleaned up
+  class NotExportedError < BaseError
+  end
+
   attr_reader :mode
 
   # Initialize the GPIO pin
@@ -57,24 +67,21 @@ class GPIO
   # @param mode [String] pin mode : IN or OUT
   def initialize(pin, mode = OUT)
     @pin = pin
-    begin
-      File.open("#{LIB_PATH}/unexport", 'w') do |file|
-        file.write(@pin)
-      end
-    rescue Errno::EINVAL
-      # Do nothing - the pin is already unexported
-    end
-    File.open("#{LIB_PATH}/export", 'w') do |file|
-      file.write(@pin)
-    end
+
+    unexport_pin
+    export_pin
+
     @mode = mode
+    @exported = true
   end
 
   # Set the pin mode
   #
   # @param mode [String] pin mode : IN or OUT
-  # @raise [UnknownMode] if the mode isn't IN or OUT
+  # @raise [NotExportedError] if the pin was already cleaned up
+  # @raise [UnknownModeError] if the mode isn't IN or OUT
   def mode=(new_mode)
+    raise NotExportedError, "gpio error : pin was already cleaned up" unless exported
     raise UnknownModeError, "gpio error : unknown mode #{new_mode}" unless PIN_MODES.include?(new_mode)
 
     File.open("#{LIB_PATH}/gpio#{@pin}/direction", 'w') do |file|
@@ -95,14 +102,40 @@ class GPIO
   #   This method can only be used when the pin is in OUT mode
   #
   # @param new_value [Integer] the value : LOW or HIGH
-  # @raise [NotOutMode] if the pin isn't in OUT mode
-  # @raise [BadValue] if the provided value isn't LOW or HIGH
+  # @raise [NotExportedError] if the pin was already cleaned up
+  # @raise [NotOutModeError] if the pin isn't in OUT mode
+  # @raise [BadValueError] if the provided value isn't LOW or HIGH
   def value=(new_value)
+    raise NotExportedError, "error : pin was already cleaned up" unless exported
     raise NotOutModeError, "error : mode isn't OUT" unless @mode == OUT
     raise BadValueError, "error : bad pin value" unless PIN_VALUE.include?(new_value)
 
     File.open("#{LIB_PATH}/gpio#{@pin}/value", 'w') do |file|
       file.write(new_value)
+    end
+  end
+
+  # Cleans up the pin by unexporting it
+  def cleanup
+    unexport_pin
+    @exported = false
+  end
+
+  private
+
+  attr_reader :exported
+
+  def unexport_pin
+    File.open("#{LIB_PATH}/unexport", 'w') do |file|
+      file.write(@pin)
+    end
+  rescue Errno::EINVAL
+    # Do nothing - the pin is already unexported
+  end
+
+  def export_pin
+    File.open("#{LIB_PATH}/export", 'w') do |file|
+      file.write(@pin)
     end
   end
 end
